# dashboard/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "description":"This is the api for users"
    },
    "tags":[
        {
            "name":"PUBLIC ROUTE",
            "description":"Public Route"
        }
    ],
    "paths":{
        "/health":{
            "summary":"Testing User Check",
            "get":{
                "description":"User Api Health Check",
                "tags":["PUblic Route"],
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        },
        "/users":{
            "summary":"Testing User Route",
            "get":{
                "description":"User Api Check",
                "tags":["PUblic Route"],
                "parameters":[
                    {
                        "in":"query",
                        "name":"role",
                        "schema":{
                            "type":"string"
                        },
                        "description":"Choose Role As Admin or User"
                    },
                    {
                        "in":"query",
                        "name":"city",
                        "schema":{
                            "type":"string"
                        },
                        "description":"Enter City Name"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        },
        "/user/{id}":{
            "summary":"Testing User Route",
            "get":{
                "description":"User Api Check",
                "tags":["PUblic Route"],
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "required":true,
                        "schema":{
                            "type":"string"
                        },
                        "description":"Choose Role As Admin or User"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        },
        "/addUser":{
            "summary":"Testing User Route",
            "post":{
                "description":"User Api Check",
                "tags":["PUblic Route"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "name": "Nikita",
                                "city": "London",
                                "phone": 987654231,
                                "role": "Admin",
                                "isActive": true
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        },
        "/updateUser":{
            "summary":"Testing User Route",
            "put":{
                "description":"User Api Check",
                "tags":["PUblic Route"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id":"",
                                "name": "Nikita",
                                "city": "London",
                                "phone": 987654231,
                                "role": "Admin",
                                "isActive": true
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        },
        "/deleteUser":{
            "summary":"Testing User Route",
            "delete":{
                "description":"User Api Check",
                "tags":["PUblic Route"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id":""
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User Health Pass"
                    }
                }
            }
        }
    }
}